backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
strawberry-graphql>=0.171.1
mcp-server>=0.1.0
starlette>=0.26.1
asyncio>=3.4.3
websockets>=10.4
python-multipart>=0.0.6 